+++ NanaLand/NanaLand/Restaurant/RestaurantDetailView.swift
@@ -266,11 +266,9 @@ struct RestaurantDetailView: View {
                                                     .foregroundStyle(Color.main)
                                                 Spacer()
                                                 ForEach(1...5, id: \.self) { number in
-                                                    Image(systemName: "star.fill")
+                                                    Image((Double(number) - viewModel.state.getReviewDataResponse.totalAvgRating) < 0 || (Double(number) - viewModel.state.getReviewDataResponse.totalAvgRating <= 0.5 && Double(number) - viewModel.state.getReviewDataResponse.totalAvgRating >= 0) ? "icStarFill" : "icStar")
                                                         .resizable()
-                                                        .aspectRatio(contentMode: .fit)
-                                                        .frame(width: 24)
-                                                        .foregroundColor((Double(number) - viewModel.state.getReviewDataResponse.totalAvgRating) < 0 || (Double(number) - viewModel.state.getReviewDataResponse.totalAvgRating <= 0.5 && Double(number) - viewModel.state.getReviewDataResponse.totalAvgRating >= 0) ? .yellow : .gray2)
+                                                        .frame(width: 17, height: 17)
                                                 }
                                                 Text("\(String(format: "%.1f" , viewModel.state.getReviewDataResponse.totalAvgRating))")
                                                     .font(.body02_bold)
@@ -304,7 +302,9 @@ struct RestaurantDetailView: View {
                                                                     Text("리뷰 \(viewModel.state.getReviewDataResponse.data[index].memberReviewCount ?? 0)")
                                                                         .font(.caption01)
                                                                     Text("ㅣ")
-                                                                    Image("icRatingStar")
+                                                                    Image("icStarFill")
+                                                                        .resizable()
+                                                                        .frame(width: 11, height: 11)
                                                                     Text("\(String(format: "%.1f", viewModel.state.getReviewDataResponse.data[index].rating ?? 0))")
                                                                         .font(.caption01)
                                                                 }
